% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italygdpindicator.R
\name{italygdp}
\alias{italygdp}
\title{Clean Corpus with BBC News article}
\usage{
italygdp()
}
\description{
Clean documents in Corpus which include BBC News article text
}
\details{
Transform the symbols like "/", "\\" into space.
         Stem text, remove numbers, remove english stopwords, extra spaces and punctuation, convert text to lower case
}
